// src/api/generated/models/ChatCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ChatCreate = {
  channel_id: number;
  external_chat_id: string;
  phone?: string;
  name?: string;
  contragent_id?: number;
  assigned_operator_id?: number;
};