// src/api/generated/models/ChatUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ChatUpdate = {
  status?: string;
  assigned_operator_id?: number;
  contragent_id?: number;
  phone?: string;
  name?: string;
  first_message_time?: string;
  first_response_time_seconds?: number;
  last_message_time?: string;
  last_response_time_seconds?: number;
};