// src/api/generated/models/MessageCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MessageCreate = {
  chat_id: number;
  sender_type: string;
  content: string;
  message_type?: string;
  status?: string;
  image_url?: string;
  source?: string;
};