// src/api/generated/models/MessageResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MessageResponse = {
  id: number;
  chat_id: number;
  sender_type: string;
  message_type: string;
  content: string;
  external_message_id?: string;
  status: string;
  created_at: string;
  updated_at: string;
  sender_avatar?: string;
  source?: string;
};