// src/api/generated/models/MessagesList.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { MessageResponse } from './MessageResponse';

export type MessagesList = {
  data: Array<MessageResponse>;
  total: number;
  skip: number;
  limit: number;
  date?: string;
};